cache: restore origin only reload
Basically reverting commit 58a9f88b8c021a5e056b883053f257f1a898adf7.
We can use origin_only  in case we are snapshot's origin,
as we do support this stack.

So when we are 'uncaching'  origin+snaps - we do need to reload only
origin and we do not need to play with snaps.